Moved completion checks to a single file for all the levels.
